lib: date_time: Refactoring
Changes in PR https://github.com/nrfconnect/sdk-nrf/pull/6125 has
caused code to become a bit more complex so there is a need for some
refactoring, which was not done in the mentioned PR to avoid risks
close to NCS 1.8.0 deadline.
