+++ lib/date_time/CMakeLists.txt
@@ -6,3 +6,6 @@
 
 zephyr_library()
 zephyr_library_sources(date_time.c)
+zephyr_library_sources(date_time_core.c)
+zephyr_library_sources_ifdef(CONFIG_DATE_TIME_NTP date_time_ntp.c)
+zephyr_library_sources_ifdef(CONFIG_DATE_TIME_MODEM date_time_modem.c)
