+++ lib/date_time/date_time_ntp.h
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2021 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#ifndef DATE_TIME_NTP_H_
+#define DATE_TIME_NTP_H_
+
+int date_time_ntp_get(int64_t *date_time_ms);
+
+#endif /* DATE_TIME_NTP_H_ */
