+++ samples/nrf9160/modem_shell/prj.conf
@@ -112,7 +112,6 @@ CONFIG_LTE_LC_MODEM_SLEEP_NOTIFICATIONS_THRESHOLD_MS=10240
 #CONFIG_LTE_NETWORK_MODE_LTE_M_GPS=y
 
 CONFIG_DATE_TIME=y
-CONFIG_DATE_TIME_UPDATE_INTERVAL_SECONDS=0
 
 # nRF Cloud
 # nRF Cloud A-GPS, disable this in case you want to enable SUPL below
