+++ tests/lib/date_time/prj.conf
@@ -12,6 +12,7 @@ CONFIG_DATE_TIME=y
 CONFIG_DATE_TIME_MODEM=n
 CONFIG_DATE_TIME_NTP=n
 CONFIG_DATE_TIME_UPDATE_INTERVAL_SECONDS=0
+CONFIG_DATE_TIME_TOO_OLD_SECONDS=0
 
 # General
 CONFIG_NEWLIB_LIBC=y
