+++ tests/lib/date_time/testcase.yaml
@@ -1,8 +1,7 @@
 tests:
   date_time.functionality_test:
-    platform_allow: nrf9160dk_nrf9160 qemu_x86 native_posix qemu_cortex_m3
+    platform_allow: nrf9160dk_nrf9160 qemu_x86 qemu_cortex_m3
     integration_platforms:
       - nrf9160dk_nrf9160
-      - native_posix
       - qemu_cortex_m3
     tags: date_time
